issue#63: refactor core tests
